# backend/docs/swagger.yaml
definitions:
...
  internal_handler.CreateOrderRequest:
    properties:
      idempotency_key:
        type: string
      items:
        items:
          properties:
            product_id:
              type: string
            quantity:
              type: integer
          required:
          - product_id
          - quantity
          type: object
        minItems: 1
        type: array
    required:
    - items
    type: object
...
  internal_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/orders:
    get:
      description: List all orders with optional status filter (admin only)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by order status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: List all orders
      tags:
      - admin
  /admin/orders/{order_no}/refund:
    post:
      description: Process a refund for an order (admin only)
      parameters:
      - description: Order number
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Refund order
      tags:
      - admin
  /admin/posts/{id}/approve:
    post:
      description: Approve a pending post (admin only)
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Approve post
      tags:
      - admin
  /admin/posts/{id}/reject:
    post:
      description: Reject a pending post (admin only)
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Reject post
      tags:
      - admin
  /admin/posts/pending:
    get:
      description: List posts awaiting admin approval
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: List pending posts
      tags:
      - admin
  /admin/products/{id}/approve:
    post:
      description: Approve a pending product (admin only)
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Approve product
      tags:
      - admin
  /admin/products/{id}/reject:
    post:
      description: Reject a pending product (admin only)
      parameters:
      - description: Product UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Reject product
      tags:
      - admin
  /admin/products/pending:
    get:
      description: List products awaiting admin approval
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: List pending products
      tags:
      - admin
  /admin/stats:
    get:
      description: Get platform dashboard statistics (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Dashboard stats
      tags:
      - admin
  /admin/users:
    get:
      description: List all users (admin only)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - admin
  /admin/users/{id}/ban:
    post:
      description: Ban a user account (admin only)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Ban user
      tags:
      - admin
...
  /orders:
    get:
      description: List orders for the authenticated user (requires auth)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by order status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: List my orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with product items (requires auth)
      parameters:
      - description: Order items
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
...
  /orders/{order_no}/refund:
    post:
      description: Process a refund for an order (requires auth)
      parameters:
      - description: Order number
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Refund an order
      tags:
      - orders
...
swagger: "2.0"